Add support for CORS origin http://localhost:3000
This commit adds support for the CORS origin http://localhost:3000 to the dev profile. This allows the frontend application running on http://localhost:3000 to make requests to the backend API. Previously, this origin was not allowed, causing cross-origin requests to fail.

The allowed-origins property in the application-dev.yml file has been updated to include http://localhost:3000. This change ensures that the frontend application can communicate with the backend API without any CORS-

This commit also removes the unused Register component, which was previously added for user registration. removed for temp. debugging